backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langgraph-checkpoint-postgres
langchain-openai
langchain-core
pydantic
asyncpg
sqlalchemy[asyncio]
python-dotenv
supabase
PyJWT
pyairtable
httpx
langchain-postgres
psycopg[binary]
pypdf
langchain-community
langchain-text-splitters
